+++ src/components/Paper/Paper.stories.tsx
@@ -1,7 +1,7 @@
 import React, { useRef } from 'react';
 import { Meta, StoryObj, StoryFn } from '@storybook/react';
 import { Paper, PaperProps } from './Paper';
-import { Stack, Figure, Image, Button, Icon } from '../index';
+import { Stack, Figure, Image, Button, Icon, Typography } from '../index';
 import { ConnectTheme } from '../ConnectTheme';
 import { GradeBand } from '../../enum/gradeband';
 import { ButtonMenuProvider } from '../../context/ButtonMenuContext';
@@ -11,7 +11,7 @@ const meta: Meta<typeof Paper> = {
   title: 'Layout/Paper',
   tags: ['autodocs'],
   parameters: {
-    layout: 'centered',
+    layout: 'fullscreen',
   },
 };
 
@@ -53,7 +53,7 @@ Aside.args = {
   ...Default.args,
   children: 'Aside',
   element: 'aside',
-  elevation: 2,
+  elevation: 4,
   className: 'connect__aside',
   fullWidth: false,
 };
@@ -70,10 +70,6 @@ FullWidth.args = {
   gradeBand: GradeBand.G4_5,
 };
 
-Aside.parameters = {
-  layout: 'fullscreen',
-};
-
 export const Card: Story = Template.bind({});
 Card.args = {
   ...Default.args,
@@ -86,14 +82,16 @@ Card.args = {
         justifyContent: 'start',
         flexWrap: `wrap`,
       }}
-      md={{
-        direction: `row`,
-        alignItems: 'center',
-        justifyContent: 'space-between',
-      }}
     >
       <Figure caption="This is a caption">
-        <Image imageSrc="" altText={'This is Alt Text'} />
+        <Image
+          imageSrc=""
+          roundedCorners={{
+            topLeft: true,
+            topRight: true,
+          }}
+          altText={'This is Alt Text'}
+        />
       </Figure>
 
       <Stack
@@ -104,13 +102,18 @@ Card.args = {
           Cancel
         </Button>
         <Button primary={true}>
-          <Icon id="check" size="sm" /> Save
+          <Icon id="check" size="sm" />
+          <Typography>Save</Typography>
         </Button>
       </Stack>
     </Stack>
   ),
   element: 'section',
-  elevation: 2,
+  elevation: 4,
   roundedCorner: true,
   className: '',
+};
+
+Card.parameters = {
+  layout: 'centered',
 };